Fix - misc
It's been a while, did an overhaul of player-prefab in order to facillitate spawning. To spawn in player press space.

Included a player spawner and a gamemaster, will use these to implement an actual game with yknow lives and stuff

pretty tired of this project

Bugs to fix: joystick stability (i dont know why im having this god awful problem of the stick flying around)